server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
# openai==0.28.0
PyPDF2==3.0.1
python-jose[cryptography]==3.3.0
# chromadb==0.4.18
# sentence-transformers==2.6.1
# numpy==1.24.3
werkzeug==2.3.7
google-generativeai==0.5.4
celery==5.6.2
redis==4.5.5